Minor upgrades for the Spinner menu on CardDetailFragment.java. Now it shows the complete names of the different sets and printings and loads new versions of the cards.
TODO: Debug the AllPrintingsFragment so when a new version of the card is clicked, that version is inflated onto CardDetailFragment

TODO: A title for the scroll down must be implemented, as well as the capability of discriminating by language before discriminating by set. Also, the names of the sets should be fixed in the scroll down menu instead of the set codes.

 ISSUE PERSISTS. App is stable BUT: The SVG views are being loaded according to the Logs but not displayed on the CardDetailFragment.java