Fixed the Spinner and inflation logic on the CardDetailFragment and AllPrintingsFragment. Correct Behavior.
TODO: The Spinner doesn't remain populated after a new version is called from AllPrintingsFragment, this has to be improved.

TODO: A title for the scroll down must be implemented, as well as the capability of discriminating by language before discriminating by set. Also, the names of the sets should be fixed in the scroll down menu instead of the set codes.

 ISSUE PERSISTS. App is stable BUT: The SVG views are being loaded according to the Logs but not displayed on the CardDetailFragment.java